+++ src/actions/index.js
@@ -38,6 +38,7 @@ export {
   getControlpanel,
   listControlpanels,
   updateControlpanel,
+  getSystemInformation,
 } from '@plone/volto/actions/controlpanels/controlpanels';
 export { getDiff } from '@plone/volto/actions/diff/diff';
 export { emailNotification } from '@plone/volto/actions/emailNotification/emailNotification';