+++ src/reducers/controlpanels/controlpanels.test.js
@@ -26,6 +26,7 @@ describe('Controlpanels reducer', () => {
       },
       controlpanel: null,
       controlpanels: [],
+      systeminformation: null,
     });
   });