+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Feature
 
 - Load the current object language on SSR, instead of relying on the cookie @sneridagh
+- Add system information in controlpanel @iFlameing #1457
 
 ### Bugfix